refactor: waitForElement
